+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Upload local images asynchronously (sc-3366)
+
 ---
 
 ## [1.3.1] - 2024-05-27