+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added upload_image_directory function for Dataset class (sc-3367)
 
+### Changed
+
+- Upload local images asynchronously (sc-3366)
+
 ---
 
 ## [1.3.1] - 2024-05-27