+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added upload_image_directory function for Dataset class (sc-3367)
+- Added INFO level logs to track the local image upload progress (sc-3368)
 
 ### Changed